test: finished all unit tests

+++ CarRental.Domain/Entities/RentalLog.cs
@@ -6,5 +6,5 @@ public class RentalLog
 	public required Car Car { get; set; }
 	public required Client Client { get; set; }
 	public required DateTime RentStartDate { get; set; }
-	public required float Duration { get; set; }
+	public required Decimal Duration { get; set; }
 }
